вопрос с прокидыванием
